# openapi.yaml
openapi: 3.0.3
info:
  title: Personal Investment Management API
  version: 0.2.0
  description: MVP API for personal investment management tool (snapshot-authoritative).
servers:
- url: '{{baseUrl}}'
  description: Variable base URL
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
        account_id:
          type: integer
        category:
          type: string
          enum:
          - buy
          - redeem_request
          - redeem_settle
          - fee
        trade_date:
          type: string
          format: date
        settle_date:
          type: string
          format: date
          nullable: true
        amount:
          type: number
          description: '买入为正，卖出为负'
        note:
          type: string
          nullable: true
      required:
      - id
      - product_id
      - account_id
      - category
      - trade_date
      - amount
...
paths:
...
  /api/products/{id}/transactions:
    get:
      summary: Get product transactions
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: window
        in: query
        required: false
        schema:
          type: string
          enum:
          - 4w
          - 8w
          - 12w
          - 24w
          - 1y
          - ytd
          default: 8w
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  data:
                    type: object
                    properties:
                      product_id:
                        type: integer
                      items:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                            account_id:
                              type: integer
                            category:
                              type: string
                              enum:
                              - buy
                              - redeem_request
                              - redeem_settle
                              - fee
                            trade_date:
                              type: string
                              format: date
                            settle_date:
                              type: string
                              format: date
                              nullable: true
                            amount:
                              type: number
                            note:
                              type: string
                              nullable: true
                  message:
                    type: string
                required:
                - code
                - data
                - message
...
  /api/transactions:
    post:
      summary: Create transaction
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  data:
                    $ref: '#/components/schemas/Transaction'
                  message:
                    type: string
                required:
                - code
                - data
                - message
              example:
                code: 200
                data:
                  id: 50001
                  product_id: 200
                  account_id: 10
                  category: buy
                  trade_date: '2025-01-05'
                  settle_date: null
                  amount: 100000
                  note: 申购28D
                message: ok
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id:
                  type: integer
                account_id:
                  type: integer
                category:
                  type: string
                  enum:
                  - buy
                  - redeem_request
                  - redeem_settle
                  - fee
                trade_date:
                  type: string
                  format: date
                settle_date:
                  type: string
                  format: date
                  nullable: true
                amount:
                  type: number
                  description: '买入为正，卖出为负'
                note:
                  type: string
                  nullable: true
              required:
              - product_id
              - account_id
              - category
              - trade_date
              - amount
            example:
              product_id: 200
              account_id: 10
              category: buy
              trade_date: '2025-01-05'
              amount: 100000
              note: 申购28D
    get:
      summary: List transactions with filters
      parameters:
      - name: product_id
        in: query
        required: false
        schema:
          type: integer
      - name: account_id
        in: query
        required: false
        schema:
          type: integer
      - name: category
        in: query
        required: false
        schema:
          type: string
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Transaction'
                      pagination:
                        type: object
                        properties:
                          page:
                            type: integer
                          page_size:
                            type: integer
                          total:
                            type: integer
                          total_pages:
                            type: integer
                  message:
                    type: string
                required:
                - code
                - data
                - message
  /api/transactions/{id}:
    delete:
      summary: Delete transaction
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                  message:
                    type: string
                required:
                - code
                - data
                - message
        '404':
          description: Not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                  message:
                    type: string
                required:
                - code
                - message
...